# Improve AVX2 encoding speed